Implement RandomTicker and configuration for it.
This commit adds a component called RandomTicker to the project. A random ticker is similar to a time.Ticker in the standard library, however it also delays its ticks by a random period of time that is within a specified range.

For example, the random ticker configured in example-config.json (as of this commit) will have a base tick interval of 5 seconds. However, the actual tick sent on the random ticker's channel will be delayed by some random amount of time between 2 and 4 seconds.

This commit also adds a debug toggle to the configuration options. This is simply a convenience so that I don't have to worry about hitting API throttle limits (or spending twilio $$) during testing.